Add preprocessor and remove content for multipart.
An `OperationPreprocessor` has been added to modify
binary content for parts, so that the snippets are not
filled with garbled data.

Also removed the body content for multipart requests
so that it isn't used for snippets. The `OperationRequestPart`s
will be used instead.

Also added an integration test for multipart.