fix: form params added to query string and update SRD version
Changes in Spring REST Docs 1.1.1 caused the form paramters to be added
to the query string. The check is based on the body content of
the OperationRequest. With MockMvc and RestAssured, the form parameters
are not added to the body content. This is not the case with this
Jersey implementation. This commit just checks to see if the request
is a forms request. If it is, then just set the content body
as an empty byte[], which is consistent with the result
of MockMvc and RestAssured.

fixes #4, closes #2